Extracting the client test property repository into its own repository since we query at it different ways
