+++ service/src/main/java/tds/config/repositories/impl/ConfigRepositoryImpl.java
@@ -66,53 +66,4 @@ public class ConfigRepositoryImpl implements ConfigRepository {
 
         return clientSystemFlags;
     }
-
-    @Override
-    public Optional<ClientTestProperty> findClientTestProperty(final String clientName, final String assessmentId) {
-        final SqlParameterSource parameters = new MapSqlParameterSource("clientName", clientName)
-                .addValue("assessmentId", assessmentId);
-
-        final String SQL =
-                "SELECT\n" +
-                "   clientname AS clientName,\n" +
-                "   testid AS assessmentId,\n" +
-                "   maxopportunities AS maxOpportunities,\n" +
-                "   prefetch AS prefetch,\n" +
-                "   isselectable AS isSelectable,\n" +
-                "   label AS label,\n" +
-                "   subjectname AS subjectName,\n" +
-                "   initialabilitybysubject AS initialAbilityBySubject,\n" +
-                "   accommodationfamily AS accommodationFamily,\n" +
-                "   sortorder AS sortOrder,\n" +
-                "   rtsformfield AS rtsFormField,\n" +
-                "   requirertswindow AS requireRtsWindow,\n" +
-                "   tide_id AS tideId,\n" +
-                "   rtsmodefield AS rtsModeField,\n" +
-                "   requirertsmode AS requireRtsMode,\n" +
-                "   requirertsmodewindow AS requireRtsModeWindow,\n" +
-                "   deleteunanswereditems AS deleteUnansweredItems,\n" +
-                "   abilityslope AS abilitySlope,\n" +
-                "   abilityintercept AS abilityIntercept,\n" +
-                "   validatecompleteness AS validateCompleteness,\n" +
-                "   gradetext AS gradeText\n" +
-                "FROM\n" +
-                "   configs.client_testproperties\n" +
-                "WHERE\n" +
-                "   clientname = :clientName\n" +
-                "   AND testid = :assessmentId";
-
-        Optional<ClientTestProperty> maybeClientTestProperty;
-        try {
-            maybeClientTestProperty = Optional.of(
-                    jdbcTemplate.queryForObject(
-                            SQL,
-                            parameters,
-                            new ClientTestPropertyRowMapper()));
-        } catch (EmptyResultDataAccessException e) {
-            LOG.debug("{} did not return results for clientName = {}, assessmentId = {}", SQL, clientName, assessmentId);
-            maybeClientTestProperty = Optional.empty();
-        }
-
-        return maybeClientTestProperty;
-    }
 }
