+++ service/src/main/java/tds/config/services/impl/ConfigServiceImpl.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tds.config.ClientSystemFlag;
 import tds.config.ClientTestProperty;
+import tds.config.repositories.ClientTestPropertyQueryRepository;
 import tds.config.repositories.ConfigRepository;
 import tds.config.services.ConfigService;
 
@@ -13,15 +14,17 @@ import java.util.Optional;
 @Service
 public class ConfigServiceImpl implements ConfigService {
     private final ConfigRepository configRepository;
+    private final ClientTestPropertyQueryRepository clientTestPropertyQueryRepository;
 
     @Autowired
-    public ConfigServiceImpl(ConfigRepository configRepository) {
+    public ConfigServiceImpl(ConfigRepository configRepository, ClientTestPropertyQueryRepository clientTestPropertyQueryRepository) {
         this.configRepository = configRepository;
+        this.clientTestPropertyQueryRepository = clientTestPropertyQueryRepository;
     }
 
     @Override
     public Optional<ClientTestProperty> findClientTestProperty(final String clientName, final String assessmentId) {
-        return configRepository.findClientTestProperty(clientName, assessmentId);
+        return clientTestPropertyQueryRepository.findClientTestProperty(clientName, assessmentId);
     }
 
     @Override
